Rename the files to make sure it matches the exact requirements
