+++ src/redux/configureStore.js
@@ -1,6 +1,6 @@
 import { combineReducers, configureStore } from 'redux';
-import bookReducer from './books/books';
-import categoriesReducer from './categories/categories';
+import bookReducer from './books/booksSlice';
+import categoriesReducer from './categories/categoriesSlice';
 
 const rootReducer = combineReducers({
   books: bookReducer,
